requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3  # Compatible with OpenCV

# Machine Learning
tensorflow==2.15.0  # For FaceNet
annoy==1.17.2       # For approximate nearest neighbor search
mediapipe==0.10.8   # For facial landmarks
face-recognition==1.3.0  # Fallback face detection
rembg==2.0.50

# Model Optimization (Optional - for production deployment)
# tf2onnx==1.15.0     # For ONNX export
# onnxruntime==1.15.1 # For ONNX inference

# Utilities
python-dotenv==1.0.0
qrcode==7.4.2

# Testing (Optional)
pytest==7.4.0
pytest-flask==1.2.0
cairosvg==2.8.2